fix(ui): move commodityDisplayName out of features for widget imports
Unblocks PR #3991: repo.app_widget_imports failed because ResourceLabelInline
imported features/, and the economic hover test still matched camelCase ids.
