+++ SPEC/program/repo-lint.md
@@ -433,7 +433,7 @@ Do **not** add new top-level `tool/check_*.dart` **entrypoints** for CI without 
 - Given the #2073 workspace outdated audit targets (repo root plus every workspace member whose `pubspec.yaml` declares `sdk: flutter`), when repo lint runs rule `repo.workspace_outdated_resolvable`, then each package row reports `current` equal to `resolvable` and the run fails listing package root/name/version deltas when any row is below `Resolvable`.
 - Given the #2073 workspace outdated audit targets (repo root plus every workspace member whose `pubspec.yaml` declares `sdk: flutter`), when repo lint runs rule `repo.workspace_outdated_latest_direct`, then every `kind: direct` or `kind: dev` row where `resolvable == latest` reports `current == latest`, and the run fails listing package root/name/version deltas when any direct row is below Latest despite Latest already being resolvable.
 - Given `packages/colonizethis_world/lib/src/**` Dart sources excluding `world/province_lookup.dart` and `world/unit_lookup.dart`, when repo lint runs rule `repo.logic_dual_region_province_field_access`, then at most 0 physical lines contain `oldWorld.provinces`, `newWorld.provinces`, `oldWorld.units`, `newWorld.units`, or manual `if (regionId == kRegionOldWorld)` / `else if (regionId == kRegionOldWorld)` branching, and the checker does not use keyed waiver tables (whole-file exclusion of canonical helper files only).
-- Given `app/lib/**/*.dart` Dart sources excluding generated suffixes (`*.g.dart`, `*.freezed.dart`, `*.mocks.dart`, `*.gen.dart`), generated app l10n paths under `packages/colonizethis_app_l10n/lib/l10n/gen/`, the Widgetbook entrypoint `app/lib/widgetbook.dart` and catalog stories under `widgetbook_host/lib/catalogs/`, the E2E expected-line snapshots under `packages/colonizethis_app_fixtures/lib/test_support/`, test paths, and any path containing a fixture / golden marker (`/test_data/`, `/fixtures/`, `/golden/`, …), when `dart run tool/ct_repo_lint.dart` runs rule `repo.app_no_duplicate_helpers`, then each tracked canonical helper symbol from #2180 (`eraRoman` and `techCategoryLabelL10n` in `app/lib/features/game/widgets/technology/tech_ui_helpers.dart`; `commodityDisplayName` in `app/lib/features/game/widgets/production/commodity_ui_helpers.dart`; the `trainDialog*` and `*TrainDialog*` set in `app/lib/features/game/widgets/train/train_unit_dialog_helper.dart`) is declared as a top-level function only in its canonical file, no removed helper (`_eraRoman`, `_categoryLabel`, `_categoryLabelL10n`, `_commodityDisplayName`, or the deleted public wrapper `unitsPanelRegionLabel` replaced by `regionDisplayLabel`, #3279) appears as a top-level function or class method anywhere in the scanned tree, the run fails listing every offending file path and start line when any tracked symbol redefinition or removed-helper reappearance is found, and the checker does not load keyed waiver tables (whole-file exclusion of generated/l10n/widgetbook/test_support/fixture paths only per this section).
+- Given `app/lib/**/*.dart` Dart sources excluding generated suffixes (`*.g.dart`, `*.freezed.dart`, `*.mocks.dart`, `*.gen.dart`), generated app l10n paths under `packages/colonizethis_app_l10n/lib/l10n/gen/`, the Widgetbook entrypoint `app/lib/widgetbook.dart` and catalog stories under `widgetbook_host/lib/catalogs/`, the E2E expected-line snapshots under `packages/colonizethis_app_fixtures/lib/test_support/`, test paths, and any path containing a fixture / golden marker (`/test_data/`, `/fixtures/`, `/golden/`, …), when `dart run tool/ct_repo_lint.dart` runs rule `repo.app_no_duplicate_helpers`, then each tracked canonical helper symbol from #2180 (`eraRoman` and `techCategoryLabelL10n` in `app/lib/features/game/widgets/technology/tech_ui_helpers.dart`; `commodityDisplayName` in `app/lib/widgets/commodity_display_name.dart`; the `trainDialog*` and `*TrainDialog*` set in `app/lib/features/game/widgets/train/train_unit_dialog_helper.dart`) is declared as a top-level function only in its canonical file, no removed helper (`_eraRoman`, `_categoryLabel`, `_categoryLabelL10n`, `_commodityDisplayName`, or the deleted public wrapper `unitsPanelRegionLabel` replaced by `regionDisplayLabel`, #3279) appears as a top-level function or class method anywhere in the scanned tree, the run fails listing every offending file path and start line when any tracked symbol redefinition or removed-helper reappearance is found, and the checker does not load keyed waiver tables (whole-file exclusion of generated/l10n/widgetbook/test_support/fixture paths only per this section).
 - Given `app/test/**/*.dart` Dart sources, when `dart run tool/ct_repo_lint.dart` runs rule `repo.app_test_no_debug_init`, then every file that invokes `getDebugInitGameResult()` is listed in the checker's in-script allowlist, and the run fails listing each violating `path:line` when a non-allowlisted file calls the helper; files that mention the symbol only in comments are not flagged (Refs #3656).
 - Given the checker's in-script allowlist for rule `repo.app_test_no_debug_init`, when `dart run tool/ct_repo_lint.dart` runs and an allowlisted entry's file is missing or no longer invokes `getDebugInitGameResult()` (AST method-invocation match, so a comment-only mention does not keep the entry alive), then the run fails listing each stale entry and instructs the author to remove it, enforcing the allowlist's shrink-only invariant (Refs #3656).
 - Given `app/lib/**/*.dart` Dart sources excluding generated suffixes (`*.g.dart`, `*.freezed.dart`, `*.mocks.dart`, `*.gen.dart`) and `_test.dart` / `/test/` paths, when `dart run tool/ct_repo_lint.dart` runs rule `repo.app_region_string_literals`, then no scanned line outside `//` / `///` comments contains a string literal whose entire content equals `oldWorld` or `newWorld`, and the run passes with zero violations (Refs #3658, SPEC/program/logic-dual-region-province-access.md).
