+++ test/check_app_no_duplicate_helpers_test.dart
@@ -26,7 +26,7 @@ void _writeCanonicalHelpers(Directory tempDir) {
   File('${tempDir.path}/app/lib/features/game/widgets/technology/tech_ui_helpers.dart')
     ..createSync(recursive: true)
     ..writeAsStringSync(_kCanonicalEraRomanContents);
-  File('${tempDir.path}/app/lib/features/game/widgets/production/commodity_ui_helpers.dart')
+  File('${tempDir.path}/app/lib/widgets/commodity_display_name.dart')
     ..createSync(recursive: true)
     ..writeAsStringSync(_kCanonicalCommodityContents);
   File(
